ife_setup_pte

Setup panel treatment effects parameters

Usage

ife_setup_pte(
  yname,
  gname,
  tname,
  idname,
  data,
  nife,
  anticipation = 0,
  alp = 0.05,
  biters = 100,
  cl = 1,
  ...
)

Arguments

yname

Name of outcome in data

gname

Name of group in data

tname

Name of time period in data

idname

Name of id in data

data

balanced panel data

nife

the number of interactive fixed effects to include in the model

anticipation

how many periods before the treatment actually takes place that it can have an effect on outcomes

alp

significance level; default is 0.05

biters

number of bootstrap iterations; default is 100

cl

number of clusters to be used when bootstrapping; default is 1

...

extra arguments