ife_subset

A function for obtaining the correct subset in a staggered treatment adoption setup with panel data. This is the subset of not yet treated observations in a particular time period and in the previous nife+1 periods where nife is the number of interactive fixed effects.

A function for obtaining the correct subset in a staggered treatment adoption setup with panel data. This is the subset of not yet treated observations in a particular time period and in the previous nife+1 periods where nife is the number of interactive fixed effects.

Usage

ife_subset(data, g, tp, nife, anticipation = 0, ...)

ife_subset(data, g, tp, nife, anticipation = 0, ...)

Arguments

data

the full dataset

g

the current group

tp

the current time period

nife

the number of interactive fixed effects

anticipation

number of periods before participating treatment that treatment can have an effect on outcomes

...

extra arguments to get the subset correct

Value

list that contains correct subset of data, n1 number of observations in this subset, and disidx a vector of the correct ids for this subset.

list that contains correct subset of data, n1 number of observations in this subset, and disidx a vector of the correct ids for this subset.