staggered_ife_attgt

Computes estimates of group-time average treatment effects in an interactive treatment effects model for untreated potential outcomes by exploiting staggered treatment adoption as in Callaway and Tsyawo (2023). This function is based on the local- differencing approach (similar to the approach proposed in Callaway and Karami). See staggered_ife_attgt2 for the main approach discussed in the paper where all pre-treatment periods are used to estimate the interactive fixed effects model.

Usage

staggered_ife_attgt(
  gt_data,
  nife = 1,
  xformla = ~1,
  anticipation = 0,
  ret_ife_regs = FALSE,
  ...
)

Arguments

gt_data

data frame that is local to the specific groups and times for which we'll be computing a treatment efect estimate.

nife

The number of interactive fixed effects. Default is 1.

xformla

Formula for which covariates to include in the model. Default is ~1.

anticipation

Number of periods that treatment is anticipated. Default is 0. This is in “periods”; e.g., code will work in time periods are equally spaced but 2 years apart. In this case, to allow for treatment anticipation of 2 year (<=> 1 period), set anticipation = 1.

ret_ife_regs

Whether or not to return the first stage ife regressions; default is FALSE.

...

extra arguments

Value

pte::attgt_if object